% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exSample_Distributions-data.R
\docType{data}
\name{exSample_Distributions}
\alias{exSample_Distributions}
\title{A data frame with simulated data}
\format{
A data frame with simulated columns:
\describe{
\item{Normal}{Numeric}
\item{Light_Right_Skew}{Numeric}
\item{Heavy_Right_Skew}{Numeric}
\item{Light_Left_Skew}{Numeric}
\item{Heavy_Left_Skew}{Numeric}
\item{Positive_Kurtosis}{Numeric}
\item{Negative_Kurtosis}{Numeric}
\item{Bimodal}{Numeric}
\item{Uniform}{Numeric}
\item{Zero_Inflated}{Numeric}
\item{Exponential}{Numeric}
}
}
\source{
Generated by \code{generate_exsample_distributions()}.
}
\usage{
data(exSample_Distributions)
}
\description{
Simulated distributions scaled to 1–10 for the exsampler demo app.
}
\keyword{datasets}
