% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{generate_exsample_distributions}
\alias{generate_exsample_distributions}
\title{Generates a dataset of example distributions}
\usage{
generate_exsample_distributions(n = 1000, seed = 123)
}
\arguments{
\item{n}{Number of samples to generate}

\item{seed}{Random seed}
}
\value{
A data.frame of simulated distributions
}
\description{
Creates a data frame with various shaped distributions scaled to 1–10.
}
