% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{scale_to_1_10}
\alias{scale_to_1_10}
\title{Rescales a numeric vector to 1–10}
\usage{
scale_to_1_10(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric vector scaled to 1–10
}
\description{
Rescales a numeric vector to 1–10
}
