% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{estimate_data_increment}
\alias{estimate_data_increment}
\title{Estimate data increment from a numeric vector}
\usage{
estimate_data_increment(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A common increment value
}
\description{
Estimate data increment from a numeric vector
}
