% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{scale_to_0_10}
\alias{scale_to_0_10}
\title{Scale to 0–10
Rescales a numeric vector to 0–10}
\usage{
scale_to_0_10(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector of same length, scaled to \verb{0–10}.
}
\description{
Scale to 0–10
Rescales a numeric vector to 0–10
}
